+++ src/MyServiceBus.Abstractions/MessageConsumeContext.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -8,6 +9,7 @@ public interface MessageConsumeContext
 {
     Guid? RequestId => null;
     Guid? CorrelationId => null;
+    IDictionary<string, object> Headers => new Dictionary<string, object>();
 
     Task RespondAsync<T>(T message, Action<ISendContext>? contextCallback = null, CancellationToken cancellationToken = default) where T : class;